Split parking lots out and limit density
This takes the meta tile (256 * 4 (meta tile size) + 18 (buffer-size)
* 2 = 1060x1060), shrinks it by 24px on an edge (to effectively create
a negative buffer-size), divides it into 96x96px chunks and selects at
most 1 parking area per chunk.

The negative buffer size prevents clips icons on meta tile boundaries.

The 96px dimension can be tuned to change the density (it should
probably be powers of 2 *or* centered on the tile to avoid gaps in the
lower right).